services/web/requirements.txt
#base
flask==1.1.4
flask-restx==0.4.0
numpy==1.19.5
pandas==1.2.4 #pandas depends on numpy, so it is installed after numpy
#search
rapidfuzz==1.4.1
#oauth
requests==2.21.0
oauthlib==3.0.1
pyOpenSSL==19.0.0
Flask-Login==0.4.1
#db
mysql-connector-python==8.0.26
#wsgi server
gunicorn==20.1.0
